*** empty log message ***

+++ src/gate/persist/TestPersist.java
@@ -198,56 +198,47 @@ public class TestPersist extends TestCase
       new Long(5), new Long(25), "dog poo irritates", Factory.newFeatureMap()
     );
 
-    // save the documents
-    doc = (Document) sds.adopt(doc,null);
-    doc2 = (Document) sds.adopt(doc2,null);
-    sds.sync(doc);
-    sds.sync(doc2);
-
     // create a corpus with the documents
     Corpus corp = Factory.newCorpus("Hamish test corpus");
     corp.add(doc);
     corp.add(doc2);
-    sds.adopt(corp,null);
-    sds.sync(corp);
+    LanguageResource persCorpus = sds.adopt(corp,null);
+    sds.sync(persCorpus);
+
 
     // read the documents back
     ArrayList lrsFromDisk = new ArrayList();
-    List types = sds.getLrTypes();
-    Iterator typesIter = types.iterator();
-    while(typesIter.hasNext()) {
-      String typeName = (String) typesIter.next();
-      List lrIds = sds.getLrIds(typeName);
+    List lrIds = sds.getLrIds("gate.corpora.SerialCorpusImpl");
 
-      Iterator idsIter = lrIds.iterator();
-      while(idsIter.hasNext()) {
-        String lrId = (String) idsIter.next();
-        FeatureMap features = Factory.newFeatureMap();
-        features.put(DataStore.DATASTORE_FEATURE_NAME, sds);
-        features.put(DataStore.LR_ID_FEATURE_NAME, lrId);
-        Resource lr = Factory.createResource(typeName, features);
-        if(lrId.startsWith("Hamish test cor")) // ensure ordering regardless of OS
-          lrsFromDisk.add(0, lr);
-        else
-          lrsFromDisk.add(lr);
-      } // for each LR ID
+    Iterator idsIter = lrIds.iterator();
+    while(idsIter.hasNext()) {
+      String lrId = (String) idsIter.next();
+      FeatureMap features = Factory.newFeatureMap();
+      features.put(DataStore.DATASTORE_FEATURE_NAME, sds);
+      features.put(DataStore.LR_ID_FEATURE_NAME, lrId);
+      Resource lr = Factory.createResource( "gate.corpora.SerialCorpusImpl",
+                                            features);
+      lrsFromDisk.add(lr);
+    } // for each LR ID
 
-    } // for each LR type
-
-//System.out.println("LRs on disk" + lrsFromDisk);
+    if (DEBUG) System.out.println("LRs on disk" + lrsFromDisk);
 
     // check that the versions we read back match the originals
-    Document diskDoc = (Document) lrsFromDisk.get(1);
-    Document diskDoc2 = (Document) lrsFromDisk.get(2);
     Corpus diskCorp = (Corpus) lrsFromDisk.get(0);
+//    Document diskDoc = (Document) lrsFromDisk.get(1);
+//    Document diskDoc2 = (Document) lrsFromDisk.get(2);
 
-    assert("doc from disk not equal to memory version", doc.equals(diskDoc));
-    assert("doc2 from disk not equal to memory version", doc2.equals(diskDoc2));
-    assert("corpus from disk not equal to mem version", corp.equals(diskCorp));
+    if (DEBUG) Out.prln("Documents in corpus: " + corp.getDocumentNames());
     assert("corp name != mem name", corp.getName().equals(diskCorp.getName()));
+    if (DEBUG) Out.prln("Memory features " + corp.getFeatures());
+    if (DEBUG) Out.prln("Disk features " + diskCorp.getFeatures());
+    assert("corp feat != mem feat",
+           corp.getFeatures().equals(diskCorp.getFeatures()));
+//    assert("doc from disk not equal to memory version", doc.equals(diskDoc));
+//    assert("doc2 from disk not equal to memory version", doc2.equals(diskDoc2));
 
     // delete the datastore
-    sds.delete();
+    if (DEBUG) sds.delete();
   } // testMultipleLrs()
 
   /** Test LR deletion */
@@ -256,6 +247,7 @@ public class TestPersist extends TestCase
     // writes the bloody thing, we need to delete it from disk before calling
     // DataStore.create
     File storageDir = File.createTempFile("TestPersist__", "__StorageDir");
+    if (DEBUG) Out.prln("Corpus stored to: " + storageDir.getAbsolutePath());
     storageDir.delete();
 
     // create and open a serial data store
